CC: Use Double instead of Float in test
Close#4

+++ controlchart/src/test/scala/io/datalayer/controlchart/ControlChartTest.scala
@@ -6,7 +6,7 @@ import org.scalatest.FunSuite
 import org.scalatest.ShouldMatchers
 
 
-case class Measure(m: Float)
+case class Measure(m: Double)
 
 
 class ControlChartTest extends FunSuite with ShouldMatchers {
